Set base style for the home page
All the elements of the page are now child elements of div.
This was so done to center all the contents of the page, so that it
is easier to read the content. The normal style requires the user to
start reading from left side and go all the way till right side, ruining
the blog/article reading experience.